passagem das funções para a nova classe
